% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-grid-blank.R
\name{facet_grid_blank}
\alias{facet_grid_blank}
\title{Facet Grid with blanks where no information avail}
\usage{
facet_grid_blank(
  rows = NULL,
  cols = NULL,
  scales = "fixed",
  space = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  drop = FALSE,
  margins = FALSE,
  facets = NULL,
  blank = TRUE
)
}
\arguments{
\item{rows, cols, scales, space, shrink, labeller, as.table, switch, margins, facets}{see documentation for \code{ggplot2::facet_grid}}

\item{drop}{Default: FALSE, all factor levels will be shown, regardless of
whether or not they appear in the data.  This differs from
\code{ggplot2::facet_grid()} which defaults to TRUE}

\item{blank}{Default TRUE.  Facets which contain no data to be plotted are
shown as completely blank spaces rather than a plot with a panel
but no geometry.  This argument only has an effect if \code{drop=FALSE}}
}
\description{
Facet Grid with blanks where no information avail
}
