% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-wrap-strict.R
\name{facet_wrap_strict}
\alias{facet_wrap_strict}
\title{Facet wrap strict}
\usage{
facet_wrap_strict(
  facets,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  drop = TRUE,
  dir = "h",
  strip.position = "top"
)
}
\arguments{
\item{facets, nrow, ncol, scales, shrink, labeller, as.table, switch, drop, dir, strip.position}{see ggplot2 docs}
}
\value{
FacetWrap oject
}
\description{
Facet wrap strict
}
